third_party/WebKit/LayoutTests/svg/dom/svgpath-negative-pathLength-expected.txt

CONSOLE ERROR: line 4: Error: <path> attribute pathLength: A negative value is not valid. ("-1")
